requirements.txt
streamlit==1.42.2
streamlit-autorefresh==1.0.1
tensorflow-cpu==2.16.1
yfinance==0.2.54
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
plotly==5.24.1